+++ tests/expected/node/react/imports.js
@@ -1,4 +1,4 @@
-import { PropTypes } from 'react';
+import PropTypes from 'prop-types';
 import DefaultName from '../stub';
 import AnotherDefault, { Baz, Qux } from '../stub';