+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.0.0
+* Updated React requirement to v15.5.
+* Updated PropTypes to use the new `prop-types` package.
+
 # 3.2.1
 * Fixed an issue with the `graphql` dependency not being listed.